1、修复NDK debug模式下编译framebuffer object的问题 2、减少一些NDK爆的warning

+++ pecker_src/pecker_framework/native/pfx_native_activity.h
@@ -107,7 +107,7 @@ public:
 			if (status)
 			{
 				PECKER_LOG_ERR("m_display_context.show_view",
-						"status = %d", status);
+						"status = %d", (int)status);
 				m_display_context.close_view();
 			}
 		}
@@ -143,7 +143,7 @@ public:
 			if (status)
 			{
 				PECKER_LOG_ERR("m_display_context.show_view",
-						"status = %d", status);
+						"status = %d", (int)status);
 				m_display_context.close_view();
 			}
 		}
